# config/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auth API",
        "version": "1.0.0",
        "description": "API for user authentication, registration, and profile management"
    },
    "servers": [
        {
            "url": "https://pharmacy-ozpa.onrender.com/api",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/product": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create a new product",
                "description": "Creates a new product with a name, actor, and PDF file.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "actor": {
                                        "type": "string"
                                    },
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error or file not uploaded"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get all products",
                "description": "Returns a paginated list of products with optional filters.",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 5
                        }
                    },
                    {
                        "name": "productName",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "totalItems": {
                                            "type": "integer"
                                        },
                                        "totalPages": {
                                            "type": "integer"
                                        },
                                        "currentPage": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/product/download": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Download a product PDF",
                "description": "Downloads the PDF file associated with a product.",
                "parameters": [
                    {
                        "name": "file",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File downloaded successfully"
                    },
                    "404": {
                        "description": "PDF file not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/product/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get a product by ID",
                "description": "Returns a single product by its ID.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/productCota": {
            "post": {
                "tags": [
                    "ProductCota"
                ],
                "summary": "Create a new productCota",
                "description": "Creates a new productCota with a name, actor, and PDF file.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "actor": {
                                        "type": "string"
                                    },
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ProductCota created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductCota"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error or file not uploaded"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "get": {
                "tags": [
                    "ProductCota"
                ],
                "summary": "Get all productCotas",
                "description": "Returns a paginated list of productCotas with optional filters.",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 5
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of productCotas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductCota"
                                            }
                                        },
                                        "totalItems": {
                                            "type": "integer"
                                        },
                                        "totalPages": {
                                            "type": "integer"
                                        },
                                        "currentPage": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/productCota/download": {
            "get": {
                "tags": [
                    "ProductCota"
                ],
                "summary": "Download a productCota PDF",
                "description": "Downloads the PDF file associated with a productCota.",
                "parameters": [
                    {
                        "name": "file",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File downloaded successfully"
                    },
                    "404": {
                        "description": "PDF file not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/productCota/{id}": {
            "get": {
                "tags": [
                    "ProductCota"
                ],
                "summary": "Get a productCota by ID",
                "description": "Returns a single productCota by its ID.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ProductCota details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductCota"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "ProductCota not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Actor": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "12345"
                    },
                    "nom": {
                        "type": "string",
                        "example": "John"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "1234567890"
                    },
                    "email": {
                        "type": "string",
                        "example": "user@example.com"
                    },
                    "willaya": {
                        "type": "string",
                        "example": "Algiers"
                    },
                    "category": {
                        "type": "string",
                        "example": "Pharmacien"
                    },
                    "dataPdf": {
                        "type": "string",
                        "example": "filename.pdf"
                    },
                    "password": {
                        "type": "string",
                        "example": "password123"
                    },
                    "status": {
                        "type": "boolean",
                        
                        "default": false
                    },
                    "logo": {
                        "type": "string",
                        "example": "logo.png"
                    },
                    "nomSociete": {
                        "type": "string",
                        "example": "My Company"
                    },
                    "subscribes": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T00:00:00Z"
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T00:00:00Z"
                    }
                }
            },
...
            "Product": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "12345"
                    },
                    "name": {
                        "type": "string",
                        "example": "Sample Product"
                    },
                    "dataPdf": {
                        "type": "string",
                        "example": "filename.pdf"
                    },
                    "actor": {
                        "$ref": "#/components/schemas/Actor"
                    },
                    "delete": {
                        "type": "boolean",
                        "example": false
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T00:00:00Z"
                    }
                }
            },
            "ProductCota": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "12345"
                    },
                    "name": {
                        "type": "string",
                        "example": "Sample ProductCota"
                    },
                    "dataPdf": {
                        "type": "string",
                        "example": "filename.pdf"
                    },
                    "actor": {
                        "$ref": "#/components/schemas/Actor"
                    },
                    "delete": {
                        "type": "boolean",
                        "example": false
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T00:00:00Z"
                    }
                }
            }
        }
    }
}